Extract node and pre renderers
